Implement video and audio duration checks to sorting
Added config options for video and audio max and min duration.

It would be inefficient to check audio and video duration during download, we can do it during sorting (if the user has sorting enabled). I will probably make a separate process, e.g. post_download_filters. That way we can process it separate from sorting and stuff.

I haven't implemented the full function for removing and checking audios yet. I also plan to add the sorting filename options to the docs because there are a lot of variables that we don't include in the docs.