fix: v4 import to pydantic config
Config is immediately validated after import, but given that most of the v4 options were `str` or `bool` (except for `proxy`) importing from v4 is actually easier that importing from v5

I had to replace `SupportedSytes` Enum with a list. It was giving me validation errors because the enum names and their values are different. Even when the validation was successful, I was getting a bunch of warnings from pydantic. We need to switch back to using regular lists for those. I will make new PR for that