fix: `File is being used by another process` error if a validation error raises when running multiple configs (#695)
* Changes the startup logger behavior by creating and immediately destroying it.

It will create the startup logger before any critical section of code where we may not have the manager ready.

It will destroy the startup logger after each of these sections but only delete the startup log file if the file is empty.

These sections are:
- The setup of the manager itself
- Calls to `manager.config_manager.change_config`

Also refactors some of the logic for the main and debug loggers

* refactor: use context manager