Feature: Add Hyperlinks to Logged Paths (#720)
* Implement hyperlinks for files in console logs

Add log folder to end of run output because many people have trouble finding it.

Also added hyperlinks for the paths. In terminals that support it, the user can click on the link to open the log folder, URLs file, or config folder.

* Implement Ruff Fixes

* Hide input file runtime stat

Hide the input file and input URL groups runtime stats if the user isn't using the input file.

* refactor: use a function to generate the hyperlink

- Use a function to generate the hyperlink
- Always show `Input File` in stats but the value will be `--links (CLI args)` when `using_input_file` is `False`

* fix: reverse `using_input_file`

- `False` by default (CLI link and retry options)
- Only change it to `True` the input file is actually used

* refactor: resolve the text in hyperlinks

* refactor: show a custom text (optional)

---------

Co-authored-by: NTFSvolume <172021377+NTFSvolume@users.noreply.github.com>