perf: resolve all base paths at startup
- Related to #812

Paths inputs for ffmpeg need to be absolute to prevent errors. A single HLS could have thousands of files and resolving each one is a blocking operation. This remove the need to do it by just resolving the base paths

Everything except log file paths because they use `log_folder` as base if they are relative.

This may break database lookup from download paths because old database entries may use relative paths

It's better to wait for a new releasse before merging this