refactor: add function for deterministic custom filenames
Adds a function to the base crawler to create a custom filename with an standard format instead of manually creating a string.

This is to eventually add a config option for the user to customize the names, ex: exclude file id, include file hash, exclude resolution, etc.

The options will intentionally be very limited. The user will only be able to control what values are include/exclude. The order is fixed and the formating of each value is fixed.

This will brake existing formating for a few crawlers. Needs to be added in a new mayor release.