perf: bypass google virus warning for large files (gDrive)
Google issues a warning when trying to download large files:

> <file> is too large for Google to scan for viruses. Would you still like to download this file

The original gdown code handles this by switch over multiple URL formats and parsing the html response to extract  a download token,

Turn out we can bypass the warning by making a POST request instead of a GET request to the content URL.

This allows us to download a file with a single request instead of the up to 4 requests required before.

None of the original gdown code is used anymore.