ci: test our min dependencies are actually valid
We have a bug right now. Since v7.4.0, we need `yarl` 1.18 or newer

However, we do not have an explicit dependency of `yarl`. We have always used the indirect dependency with `aiohttp` but `aiohttp` only requires `yarl` 1.17.

This workflow will allow to catch those kind of things.

It can only check the latest public release unless we replace `poetry` with `uv`.  Not that useful right now but at least we can run it manually.