Replace WeakValueDictionary with our own file lock ref counting
I was getting a `KeyError` on `async with self._locked_files[filename]` because I think `WeakValueDictionary` was immediately allowing the lock to be garbage collected due to nothing outside it having a reference to the lock at the moment it was added.

So this manually counts the references and deletes it from the dict when it goes to 0. I'm not a Python expert though, so there might be a better way to do this.