refactor: save the database path as an attribute of `MediaItem` (#1354)
We use the db path for every db lookup. We compute it every time before a lookup but this path will never changes after the item is created

This will allow us to simplify all db calls in the future since the only argument they will ever need is the MediaItem object.