fix: check free space before downloading any chunk
We have a check before this one but, if we are downloading a high number of files, the current file may get stuck waiting for a turn in the limiter.

We need to check frre space again right before the actual download begins. However, we were downloading 5 chunks and _then_ checking again.

This PR changes it to make sure we check again before any chunk download actually begins.