fix: Hash Table Query Times
Fix extremely slow hash table querying. It seems that coomer is exceptionally slow when it comes to this for some reason.

Without this index, a coomer run by another user ran for more than 72 hours and did not download anything.
We cancelled that run, created this index, and then the same run completed in a little over 3 hours.